refactor: refactor handler
