+++ src/lib.rs
@@ -1,2 +1,4 @@
 pub mod config;
+pub mod handler;
 pub mod parsers;
+pub mod vars;
